+++ src/resources/assets/scripts/routes/createReport.js
