+++ src/resources/assets/scripts/routes/index.js
@@ -1,9 +1,9 @@
 import common from './common';
-import report from './report';
+import createReport from './createReport';
 import signup from './signup';
 
 export default {
   common,
-  createReport: report,
+  createReport,
   signup,
 };
